requirements.txt
-e ./transformers
torch
pillow
accelerate
line-profiler
torchvision
sentencepiece
pandas
tabulate